First Game, Number Guessing Game
The Number Guessing Game is able to work correctly now. I figured out that the error not allowing it to work was that I wasn't getting the entry input when I used self.guess = int(self.sVar.get()), so I had to use int(self.guessEntry.get()) instead. Then, I had to fix self.trial because it wasn't decrementing after each entry, so I had to remove it from the main function that runs the game and keep it in the previous functions.