First three games finished
I've finished the first three games: Number Guessing Game, Rock, Paper, Scissors, and the Fun Fact Game. With the Rock, Paper, Scissors Game, I added three buttons: Single Player, Multiplayer, and Back to Menu. The single-player game allows the user to player against the computer. The multiplayer game allows there to be an entry for players 1 and 2's names, and then they'll be navigated to the actual game page where they can play the game. There's only one issue with the fun fact quiz game: the messagebox that is supposed to show the result sets the number of correct answers at 0 and the number of wrong answers at 11, the total number of questions.